fix inferred any warning

+++ frontend/src/main/scala/Main.scala
@@ -31,7 +31,7 @@ object Main {
     import window.location
     val protocol = if (location.protocol == "https:") "wss" else "ws"
     //TODO: proxy with webpack devserver and only configure production port
-    val port = Config.wsPort getOrElse location.port
+    val port = Config.wsPort getOrElse location.port.toInt
 
     val apiEventHandler = createHandler[Seq[ApiEvent]]().unsafeRunSync()
     val state = new GlobalState(apiEventHandler)
